Move creation of discount codes into helper file
