+++ tests/basket.test.js
@@ -1,31 +1,7 @@
 const Basket = require("../Basket");
 const Product = require("../Product");
-const Discount = require("../discountCodes");
 
-// Fake API data
-const { getDiscountCodes } = require("../api");
-
-let discountCodes = getDiscountCodes();
-
-discountCodes = discountCodes.map(discountCode => {
-  switch (discountCode.type) {
-    case "bxgx":
-      return new Discount.Bxgx(
-        discountCode.code,
-        discountCode.productCode,
-        discountCode.options
-      );
-    case "bulkBuy":
-      return new Discount.BulkBuy(
-        discountCode.code,
-        discountCode.productCode,
-        discountCode.options
-      );
-    default:
-      console.log(new Error("An unknown discount type was found"));
-      return;
-  }
-});
+const { discountCodes } = require("../helper");
 
 test("adds 2 fruit tea, 1 strawberries and 1 coffee", () => {
   const products = [
